requirements.txt
attr
DALL-E
timm
numpy
opencv_python
scikit-learn
scipy
matplotlib
plyfile
nvidia-ml-py
yacs
tqdm
tensorboardX
argparse
sacred
progressbar2
path.py
h5py
blessings
scikit-image
lz4
lmdb
fitlog
appdirs
cachetools
certifi
cffi
chardet
cycler
Cython
decorator
easydict
fire
google-auth
google-auth-oauthlib
termcolor
got10k
grpcio
idna
imgaug
resnest
wandb
imageio
einops
fire
pillow-simd
pytorchvideo
decord
pims
moviepy
seaborn
mediapy
terminaltables